巨大ファイル編集（B4 統合）: PieceTableViewer に一本化・検索/追従を復活
大ファイルの既定を PieceTableViewer にし、MREDITOR_PIECETABLE デバッグ分岐と
LargeFileViewer を撤去。検索・フィルタ・追従・行ジャンプを piece table 上に復活。

- 表示ソースを dirty で切替: クリーン=LineIndex+mmap（追従で伸びる）／編集後=piece table
- 検索(複数語AND/正規表現/大小/ハイライト)・フィルタ表示・find next/prev を移植し
  SearchEngine(mmap) を再利用。未保存の編集がある間は無効化（mmap と乖離するため）
- tail -f を移植（クリーン時のみ・LineIndex.extend + remap）
- supportsSearch/Follow を !isDirty 連動に。編集開始で検索/追従/フィルタを停止
- ViewerState を DocumentPane.swift へ、フォント管理を EditorFont へ切り出し
- makePane: 大ファイル→PieceTableViewer（env フラグ不要）。読み取り専用バナーは
  canEdit=true で自然に非表示
- 全 36 テスト green
